Added node_modules to gitignore
This does not revert the previous fix. The addition to the gitignore only affects the top directory, as a node_modules folder is created when using checkout from master branch to executable. The node_modules folder that is required for this branch to work is in resources/app/node_modules, and is not affected by this change.